requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
black==19.10b0
blessings==1.7
bpython==0.18
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
curtsies==0.3.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
greenlet==0.4.15
httpie==1.0.3
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pathspec==0.6.0
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.4
pytest==5.2.2
python-dotenv==0.10.3
regex==2019.11.1
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.10
toml==0.10.0
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
